do not multiple cast

+++ utils/utils.py
@@ -45,12 +45,12 @@ def get_feat(batch_graph, varvar, maxclause, maxvar,dtype=torch.half):
     for ssm in batch_graph:
         if not varvar: #ie clause x var
             nclause.append(ssm.shape[0])
-            clause_arities.append(torch.cat([torch.tensor(np.asarray(np.absolute(ssm).sum(axis=1).flatten())[0]),torch.zeros(maxclause-nclause[-1],dtype=torch.int32)]))
+            clause_arities.append(torch.cat([torch.tensor(np.asarray(np.absolute(ssm).sum(axis=1).flatten())[0]),torch.zeros(maxclause-nclause[-1],dtype=dtype)]))
         nvar.append(ssm.shape[1])
         ar = torch.tensor(np.asarray(ssm.sum(axis=0).flatten())[0])
         if varvar: # we store on disk only directed links
             ar += torch.tensor(np.asarray(np.absolute(ssm).sum(axis=1).flatten())[0])
-        var_arities.append(torch.cat([ar,torch.zeros(maxvar-nvar[-1],dtype=torch.int32)]))
+        var_arities.append(torch.cat([ar,torch.zeros(maxvar-nvar[-1],dtype=dtype)]))
 
     if not varvar:
         clause_feat = torch.cat(clause_arities, 0)
@@ -60,8 +60,8 @@ def get_feat(batch_graph, varvar, maxclause, maxvar,dtype=torch.half):
     var_feat.unsqueeze_(1)
 
     if not varvar:
-        return clause_feat.to(dtype), var_feat.to(dtype), nclause, nvar
-    return None, var_feat.to(dtype), None, nvar
+        return clause_feat, var_feat, nclause, nvar
+    return None, var_feat, None, nvar
 
 
 def postproc(data, maxclause,maxvar,varvar, graph_pool=False, neg_as_link = True):
